Fixed User routes and added more verification
