+++ server.js
@@ -24,9 +24,9 @@ app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
 
 // Routes
-app.use('/v1/auth', require('./routes/api/authRoutes'));
-app.use('/v1', require('./routes/api/userRoutes'));
-app.use('/v1', require('./routes/api/categoryRoutes'));
+app.use('/v1/auth', require('./routes/api/authRoutes')); // Routes for user to register, login, and logout
+app.use('/v1/user', require('./routes/api/userRoutes'));
+app.use('/v1/categories', require('./routes/api/categoryRoutes'));
 
 mongoose.connection.once('open', () => {
     console.log('MongoDB connected!');
